Move metric wrappers to a module of their own (with some other changes to allow for that)
