+++ tests/test_multiprocess.py
@@ -12,7 +12,7 @@ if sys.version_info < (2, 7):
 else:
     import unittest
 
-from prometheus_client import core
+from prometheus_client import metric_wrappers
 from prometheus_client.core import (
     CollectorRegistry,
     Counter,
@@ -26,24 +26,29 @@ from prometheus_client.multiprocess import (
     MultiProcessCollector,
     _MultiProcessValue,
 )
+from prometheus_client.value import _MutexValue
 
 
 class TestMultiProcess(unittest.TestCase):
+
+    def _set_value_class(self, new_value_class):
+        metric_wrappers._ValueClass = new_value_class
+
     def setUp(self):
         self.tempdir = tempfile.mkdtemp()
         os.environ['prometheus_multiproc_dir'] = self.tempdir
-        core._ValueClass = _MultiProcessValue(lambda: 123)
+        self._set_value_class(_MultiProcessValue(lambda: 123))
         self.registry = CollectorRegistry()
         self.collector = MultiProcessCollector(self.registry, self.tempdir)
 
     def tearDown(self):
         del os.environ['prometheus_multiproc_dir']
         shutil.rmtree(self.tempdir)
-        core._ValueClass = core._MutexValue
+        self._set_value_class(_MutexValue)
 
     def test_counter_adds(self):
         c1 = Counter('c', 'help', registry=None)
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         c2 = Counter('c', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('c_total'))
         c1.inc(1)
@@ -52,7 +57,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_summary_adds(self):
         s1 = Summary('s', 'help', registry=None)
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         s2 = Summary('s', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('s_count'))
         self.assertEqual(0, self.registry.get_sample_value('s_sum'))
@@ -63,7 +68,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_histogram_adds(self):
         h1 = Histogram('h', 'help', registry=None)
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         h2 = Histogram('h', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('h_count'))
         self.assertEqual(0, self.registry.get_sample_value('h_sum'))
@@ -76,7 +81,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_all(self):
         g1 = Gauge('g', 'help', registry=None)
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         g2 = Gauge('g', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '123'}))
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '456'}))
@@ -88,7 +93,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_liveall(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='liveall')
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='liveall')
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '123'}))
         self.assertEqual(0, self.registry.get_sample_value('g', {'pid': '456'}))
@@ -102,7 +107,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_min(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='min')
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='min')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -111,7 +116,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_max(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='max')
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='max')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -120,7 +125,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_gauge_livesum(self):
         g1 = Gauge('g', 'help', registry=None, multiprocess_mode='livesum')
-        core._ValueClass = _MultiProcessValue(lambda: 456)
+        self._set_value_class(_MultiProcessValue(lambda: 456))
         g2 = Gauge('g', 'help', registry=None, multiprocess_mode='livesum')
         self.assertEqual(0, self.registry.get_sample_value('g'))
         g1.set(1)
@@ -136,7 +141,7 @@ class TestMultiProcess(unittest.TestCase):
 
     def test_counter_across_forks(self):
         pid = 0
-        core._ValueClass = _MultiProcessValue(lambda: pid)
+        self._set_value_class(_MultiProcessValue(lambda: pid))
         c1 = Counter('c', 'help', registry=None)
         self.assertEqual(0, self.registry.get_sample_value('c_total'))
         c1.inc(1)
@@ -149,7 +154,7 @@ class TestMultiProcess(unittest.TestCase):
     @unittest.skipIf(sys.version_info < (2, 7), "Test requires Python 2.7+.")
     def test_collect(self):
         pid = 0
-        core._ValueClass = _MultiProcessValue(lambda: pid)
+        self._set_value_class(_MultiProcessValue(lambda: pid))
         labels = dict((i, i) for i in 'abcd')
 
         def add_label(key, value):
@@ -210,7 +215,7 @@ class TestMultiProcess(unittest.TestCase):
     @unittest.skipIf(sys.version_info < (2, 7), "Test requires Python 2.7+.")
     def test_merge_no_accumulate(self):
         pid = 0
-        core._ValueClass = _MultiProcessValue(lambda: pid)
+        self._set_value_class(_MultiProcessValue(lambda: pid))
         labels = dict((i, i) for i in 'abcd')
 
         def add_label(key, value):
